+++ tests/test_submission.py
@@ -316,16 +316,6 @@ class OAuthSubmissionTest(OAuthPRAWTest):
         self.assertFalse(submission.refresh().hidden)
 
     @betamax()
-    def test_raise_invalidsubmission_oauth(self):
-        self.r.refresh_access_information(self.refresh_token['submit'])
-        submission = self.r.get_submission(
-            submission_id=self.submission_deleted_id)
-        self.assertRaises(errors.InvalidSubmission, submission.add_comment,
-                          'test')
-        invalid_post = errors.InvalidSubmission()
-        self.assertEqual(invalid_post.ERROR_TYPE, str(invalid_post))
-
-    @betamax()
     def test_remove_oauth(self):
         self.r.refresh_access_information(self.refresh_token['modposts'])
         Submission.from_id(self.r, self.submission_edit_id).remove()
@@ -335,9 +325,3 @@ class OAuthSubmissionTest(OAuthPRAWTest):
         self.r.refresh_access_information(self.refresh_token['submit'])
         result = self.r.submit(self.sr, 'OAuth Submit', text='Foo')
         self.assertTrue(isinstance(result, Submission))
-
-    @betamax()
-    def test_vote_oauth(self):
-        self.r.refresh_access_information(self.refresh_token['vote'])
-        submission = Submission.from_id(self.r, self.submission_edit_id)
-        submission.clear_vote()